libgeda: Config API improvements.
Suggested by istankovic:

- Clarified that the contexts returned by the configuration context
  accessor functions (e.g. eda_config_get_system_context()) shouldn't
  be disposed of with g_object_unref(), since they are owned by the
  library.
- Made it clearer when `length' arguments are mandatory.
- Changed the `length' arguments to eda_config_get_groups() and
  eda_config_get_keys() to gsize pointers instead of gint pointers,
  and don't attempt to dereference the `length' argument if it's NULL.
- Correctly initialised and reset `parent_handler_id' field.
- Cleaned up unused variable in eda_config_get_user_context.
- Corrected some doc comment errors.
- Update propagate_key_file_error() to cope better with bad
  destination error pointers.

+++ libgeda/include/libgeda/edaconfig.h
@@ -112,9 +112,9 @@ gboolean eda_config_is_trusted (EdaConfig *cfg);
 void eda_config_set_trusted (EdaConfig *cfg, gboolean trusted);
 EdaConfig *eda_config_get_trusted_context (EdaConfig *cfg);
 
-gchar **eda_config_get_groups (EdaConfig *cfg, gint *length) G_GNUC_WARN_UNUSED_RESULT;
+gchar **eda_config_get_groups (EdaConfig *cfg, gsize *length) G_GNUC_WARN_UNUSED_RESULT;
 gboolean eda_config_has_group (EdaConfig *cfg, const gchar *group);
-gchar **eda_config_get_keys (EdaConfig *cfg, const gchar *group, gint *length, GError **error) G_GNUC_WARN_UNUSED_RESULT;
+gchar **eda_config_get_keys (EdaConfig *cfg, const gchar *group, gsize *length, GError **error) G_GNUC_WARN_UNUSED_RESULT;
 gboolean eda_config_has_key (EdaConfig *cfg, const gchar *group, const gchar *key, GError **err);
 
 gboolean eda_config_is_inherited (EdaConfig *cfg, const gchar *group, const gchar *key, GError **err);
